+++ examples/addr2ln_pid.rs
@@ -91,7 +91,7 @@ print its symbol, the file name of the source, and the line number.",
                     print_frame(&frame.name, None, &frame.code_info);
                 }
             }
-            Symbolized::Unknown => {
+            Symbolized::Unknown(..) => {
                 println!("{input_addr:#0width$x}: <no-symbol>", width = ADDR_WIDTH)
             }
         }