+++ examples/backtrace.rs
@@ -89,7 +89,7 @@ fn symbolize_current_bt() {
                     print_frame(&frame.name, None, &frame.code_info);
                 }
             }
-            Symbolized::Unknown => {
+            Symbolized::Unknown(..) => {
                 println!("{input_addr:#0width$x}: <no-symbol>", width = ADDR_WIDTH)
             }
         }